+++ go/example_utils.go
@@ -64,11 +64,13 @@ func getExampleClient() *Client {
 		initFlags()
 	})
 	config := config.NewClientConfiguration().
-		WithAddress(&standaloneAddresses[0])
+		WithAddress(&standaloneAddresses[0]).
+		WithRequestTimeout(5 * time.Second)
 
 	client, err := NewClient(config)
 	if err != nil {
 		fmt.Println("error connecting to server: ", err)
+		return nil
 	}
 
 	standaloneClients = append(standaloneClients, client)
@@ -93,6 +95,7 @@ func getExampleClusterClient() *ClusterClient {
 	client, err := NewClusterClient(cConfig)
 	if err != nil {
 		fmt.Println("error connecting to server: ", err)
+		return nil
 	}
 
 	clusterClients = append(clusterClients, client)
@@ -123,6 +126,7 @@ func getExampleClientWithSubscription(mode config.PubSubChannelMode, channelOrPa
 	client, err := NewClient(config)
 	if err != nil {
 		fmt.Println("error connecting to server: ", err)
+		return nil
 	}
 
 	standaloneClients = append(standaloneClients, client)
@@ -153,6 +157,7 @@ func getExampleClusterClientWithSubscription(
 	client, err := NewClusterClient(ccConfig)
 	if err != nil {
 		fmt.Println("error connecting to server: ", err)
+		return nil
 	}
 
 	clusterClients = append(clusterClients, client)