+++ node/src/Commands.ts
@@ -436,6 +436,43 @@ export function createClientId(): command_request.Command {
 }
 
 /**
+ * Defines the pause mode for {@link GlideClient.clientPause} and
+ *      {@link GlideClusterClient.clientPause} commands.
+ *
+ * @see {@link https://valkey.io/commands/client-pause/|valkey.io} for details.
+ */
+export enum ClientPauseMode {
+    /** Pause all client commands. */
+    ALL = "ALL",
+
+    /** Pause client write commands. */
+    WRITE = "WRITE",
+}
+
+/**
+ * @internal
+ */
+export function createClientPause(
+    timeout: number,
+    mode?: ClientPauseMode,
+): command_request.Command {
+    const args: string[] = [timeout.toString()];
+
+    if (mode !== undefined) {
+        args.push(mode);
+    }
+
+    return createCommand(RequestType.ClientPause, args);
+}
+
+/**
+ * @internal
+ */
+export function createClientUnpause(): command_request.Command {
+    return createCommand(RequestType.ClientUnpause, []);
+}
+
+/**
  * @internal
  */
 export function createConfigGet(parameters: string[]): command_request.Command {