+++ python/glide-shared/glide_shared/commands/core_options.py
@@ -346,7 +346,7 @@ class FlushMode(Enum):
 
     `FLUSHALL` command and `FUNCTION FLUSH` command.
 
-    See [FLUSHAL](https://valkey.io/commands/flushall/) and [FUNCTION-FLUSH](https://valkey.io/commands/function-flush/)
+    See [FLUSHALL](https://valkey.io/commands/flushall/) and [FUNCTION-FLUSH](https://valkey.io/commands/function-flush/)
     for details
 
     SYNC was introduced in version 6.2.0.
@@ -372,6 +372,20 @@ class FunctionRestorePolicy(Enum):
     Appends the restored libraries to the existing libraries, replacing any existing ones in case
     of name collisions. Note that this policy doesn't prevent function name collisions, only libraries.
     """
+
+
+class ClientPauseMode(Enum):
+    """
+    Mode option for the `CLIENT PAUSE` command.
+
+    See [CLIENT PAUSE](https://valkey.io/commands/client-pause/) for details.
+    """
+
+    ALL = "ALL"
+    """ Pause all client commands. """
+
+    WRITE = "WRITE"
+    """ Pause client write commands. """
 
 
 def _build_sort_args(